requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.2.0
lxml==4.5.2
macholib==1.14
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
PyExecJS==1.5.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
